Creada la clase UsuariosDAO Creado el método comprobarUsuario en la clase UsuariosDAO Creado el método borrarUsuario en la clase UsuariosDAO Establecido un System.out.println en la clase conectarDBA para informar de que se ha cerrado la conexión. El método comprobarUsuario recibe un dni por parámetro y comprueba si existe. El método borrarUsuario recibe un dni por parámetro, lo comprueba con comprobarUsuario y si existe lo borra, en caso contrario informa al usuario de que no existe.

+++ trunk/OpenGis/src/code/google/com/opengis/gestionDAO/UsuariosDAO.java
@@ -1,20 +1,77 @@
+/*****************************************************************************
+* Copyright (c) 2011 [OpenGisTeam]                                           *
+******************************************************************************/
 package code.google.com.opengis.gestionDAO;
+import java.sql.*;
 
-import java.sql.SQLException;
+/**
+* @author Pipepito
+* Clase que establece conexión con la base de datos y nos permite insertar, modificar y eliminar usuarios.
+* 
+* Última modificación 12/11/11
+* 
+*/
+
 
 public class UsuariosDAO {
 
+static boolean existe;
+static String resultado;
+static ConectarDBA dba = new ConectarDBA();
 	
-	
-	
-	public static void borrarUsuario(String dni) throws SQLException{
+/**
+ * Este método compara el dni que le pasamos por parámetro con todos los dni de la tabla usuarios, se puede emplear para todos los métodos de la clase UsuariosDAO
+ * *****************************************IMPORTANTE****************************************
+ * *				El método abre una conexión, pero la debemos cerrar desde				 *
+ * *							el resto de métodos que llamen a este.						 *
+ * *******************************************************************************************
+ * @param dni parámetro DNI que es el campo a buscar en la tabla usuarios, si existe, cambiará el estado del boolean a true, en caso contrario a false.
+ * @throws SQLException
+ */
+public static void comprobarUsuario(String dni) throws SQLException{
 		
-		ConectarDBA dba = new ConectarDBA();
+		
 		dba.acceder();
-		String sentencia = "DELETE FROM `dai2opengis`.`usuario` WHERE `usuario`.`dni` = \'"+dni+"\'";
-		dba.modificar(sentencia);
+		String sentencia = "SELECT * FROM `usuario` WHERE `dni` LIKE '"+dni+"'";
+		ResultSet rs = dba.consulta(sentencia);
+		while(rs.next()){
+			System.out.println("Ejecuto el while");
+			resultado = rs.getString(1);
+
+		}
+			
+			System.out.println("Enviado: "+dni+" esperado: "+resultado);
+			if(resultado == null){
+				existe = false;
+				System.out.println("El estado de existe es: "+existe);
+			}else if (resultado.equals(dni)){
+				existe = true;
+				System.out.println("El estado de existe es: "+existe);
+			}
+		
+		rs.close();
+		
+		//System.out.println("Ejecutada sentencia "+sentencia);
+		
+	}
+	/** 
+	 * Este método compara un dni que obtiene por parámetro utilizando el método comprobarUsuario(), si existe, lo borrará de la BD.
+	 * Después cierra la conexión que abre el método comprobarUsuario().
+	 * @param dni Parámetro DNI del usuario que vamos a borrar.
+	 * @throws SQLException
+	 */
+	public static void borrarUsuario(String dni) throws SQLException{
+		comprobarUsuario(dni);
+		if (existe == true){ 
+			
+			
+			String sentencia = "DELETE FROM `dai2opengis`.`usuario` WHERE `usuario`.`dni` = \'"+dni+"\'";
+			dba.modificar(sentencia);
+			//System.out.println("Ejecutada sentencia:"+sentencia);
+		}else if (existe == false){
+			System.out.println("El usuario no existe");
+		}
 		dba.cerrarCon();
-		System.out.println("Ejecutada sentencia:"+sentencia);
 		
 		
 	}
